Wrote a script to parse the docstrings of a file to generate a markdown documentation file
The script is in addon/pyUtils/markdown_from_docstrings.py and has been used to parse the thermo class, to generate thermo_methods.py.
Did some small rewrites of the docstrings in thermo.py to make them consistent enough to be easily parsed.

The thermo_methods.md has also been copied to the wiki under the 'Methods in the thermo class' header.