Initial commit for prototype / proof-of-concept for PyBind11/Fortwrap wrapping
The current `pybind11_bindings.cpp` contains minor manual tweaks that are necessary and have not yet been automated. Specifically: The lambdas exposed to python take `std::vector`s as arguments (such that python can pass a `list`), and forwards `std::vector.data()` to the internal C++ method. This is necessary for non-modifiable, obligatory arguments that are arrays."