Set up automated wheel building and unittesting for macOS and (many)linux
Modified map_platform_specifics.py such that it can be run on its own (because we don't need makescript.py when using the cmake-based build).

The unittests.yml workflow builds and runs the pFUnit-based unittests and the pyTests. Currently, the test "test_critical_suite:test_critical_point_mix" fails on linux.

The cibuildwheels.yml workflow builds wheels for macOS x86_64 and arm64, and a manylinux2014 wheel, and runs the pyTests using the wheel for macOS arm64.