Fix quantum cubic critical parameters for normal hydrogen and neon.
These critical temperatures are the ones actually used
in the development of the quantum cubic EoS, and correspond to the most accurate up-to-date estimates. Regrettably, they are not equal to the critical temperatures listed in the quantum cubic article.

For completeness we also updated critical pressures and molar masses, although these don't affect the accuracy of the EoS.